dev: made findComponentMeta generic
